Expose encryption_algorithm option
We have a default aes-256-cbc algorithm in API (although I think a lot
of the columns have been converted to use GCM at this point), but
Sweetness uses a different algorithm (the default aes-256-gcm). This
means they can't talk to each other, which is problematic for obvious
reasons.

This fixes that by letting the caller specify the encryption algorithm,
thus ensuring that whatever API uses as its default is overridden.